+++ modules/commons/src/main/java/org/apache/ignite/internal/util/tostring/SBLimitedLength.java
